move files to a new folder
