+++ frontend/recepies.generated.online/package.json
@@ -13,6 +13,7 @@
     "express": "^4.17.1",
     "fibers": "^5.0.0",
     "firebase": "^8.1.1",
+    "html2canvas": "^1.0.0-rc.7",
     "instantsearch.css": "^7.4.5",
     "vue": "^2.6.12",
     "vue-cookies": "^1.7.4",
