+++ frontend/recepies.generated.online/src/components/postcard-stuff/MoneyBar.vue
